# [Refactor]: improve performance for detecting function components