--- a/lib/util/jsx.js
+++ b/lib/util/jsx.js
@@ -4,7 +4,6 @@
 
 'use strict';
 
-const estraverse = require('estraverse');
 const elementType = require('jsx-ast-utils/elementType');
 
 const astUtil = require('./ast');
@@ -95,63 +94,48 @@ function isWhiteSpaces(value) {
  * @returns {Boolean} True if the node is returning JSX or null, false if not
  */
 function isReturningJSX(isCreateElement, ASTnode, context, strict, ignoreNull) {
-  let found = false;
-  astUtil.traverseReturns(ASTnode, context, (node) => {
-    // Traverse return statement
-    astUtil.traverse(node, {
-      enter(childNode) {
-        const setFound = () => {
-          found = true;
-          this.skip();
-        };
-        switch (childNode.type) {
-          case 'FunctionExpression':
-          case 'FunctionDeclaration':
-          case 'ArrowFunctionExpression':
-            // Do not traverse into inner function definitions
-            return this.skip();
-          case 'ConditionalExpression':
-            if (!strict) break;
-            if (isJSX(childNode.consequent) && isJSX(childNode.alternate)) {
-              setFound();
-            }
-            this.skip();
-            break;
-          case 'LogicalExpression':
-            if (!strict) break;
-            if (isJSX(childNode.left) && isJSX(childNode.right)) {
-              setFound();
-            }
-            this.skip();
-            break;
-          case 'JSXElement':
-          case 'JSXFragment':
-            setFound();
-            break;
-          case 'CallExpression':
-            if (isCreateElement(childNode)) {
-              setFound();
-            }
-            this.skip();
-            break;
-          case 'Literal':
-            if (!ignoreNull && childNode.value === null) {
-              setFound();
-            }
-            break;
-          case 'Identifier': {
-            const variable = variableUtil.findVariableByName(context, childNode.name);
-            if (isJSX(variable)) {
-              setFound();
-            }
-            break;
-          }
-          default:
+  const isJSXValue = (node) => {
+    if (!node) {
+      return false;
+    }
+    switch (node.type) {
+      case 'ConditionalExpression':
+        if (strict) {
+          return isJSXValue(node.consequent) && isJSXValue(node.alternate);
         }
-      },
-    });
+        return isJSXValue(node.consequent) || isJSXValue(node.alternate);
+      case 'LogicalExpression':
+        if (strict) {
+          return isJSXValue(node.left) && isJSXValue(node.right);
+        }
+        return isJSXValue(node.left) || isJSXValue(node.right);
+      case 'SequenceExpression':
+        return isJSXValue(node.expressions[node.expressions.length - 1]);
+      case 'JSXElement':
+      case 'JSXFragment':
+        return true;
+      case 'CallExpression':
+        return isCreateElement(node);
+      case 'Literal':
+        if (!ignoreNull && node.value === null) {
+          return true;
+        }
+        return false;
+      case 'Identifier': {
+        const variable = variableUtil.findVariableByName(context, node.name);
+        return isJSX(variable);
+      }
+      default:
+        return false;
+    }
+  };
 
-    return found && estraverse.VisitorOption.Break;
+  let found = false;
+  astUtil.traverseReturns(ASTnode, context, (node, breakTraverse) => {
+    if (isJSXValue(node)) {
+      found = true;
+      breakTraverse();
+    }
   });
 
   return found;