@golopot 2022-04-06
The gist is that `isReturningJSX` tries to find return statements that contains JSX. For that a tree traversing is performed. Previously the tree traversing is inefficient because it make unnecessary traversals to nodes that cannot contain a return statement. This PR makes the tree traversal more efficient by only traverse what is needed.